backend/jobs/sales_tracker/requirements.txt
# Sales Tracker Job Dependencies
supabase>=2.0.0
pydantic>=2.4.2
requests>=2.31.0
python-dateutil>=2.8.2
certifi>=2023.11.17
python-dotenv>=1.0.0
